+++ src/index.js
@@ -4,8 +4,16 @@ import runAllTest from './run-all-tests';
 import { options } from './config';
 import { registerRequireHook } from './require-hook';
 import coverage from './coverage';
+import { getCurrentTest, events } from './context';
+// console._log_mm = console.log;
+// console.log = function (...args) {
+//     let err;
+//     console.trace();
+//     this._log_mm(...args);
+// };
 
 
+const cwd = process.cwd();
 const plugins = [];
 const availablePlugins = [
     'mokamok-react'
@@ -22,13 +30,42 @@ for (let i = 0; i < availablePlugins.length; i++) {
 }
 
 
+if (options.babelPresets.length === 0) {
+    // pick the latest available preset
+    const names = [ 'latest' ];
+    let year = new Date().getFullYear() + 2;
+    while (year > 2015) {
+        names.push(year.toString());
+        year--;
+    }
+    for (var i = 0; i < names.length; i++) {
+        try {
+            const name = names[i];
+            require(`babel-preset-${name}`);
+            options.babelPresets.push(name);
+            break;
+        } catch (err) {
+            // NOOP
+        }
+    }
+}
+
+
+function resolveBabelExt(type, path) {
+    if (path[0] === '.' || path[0] === '/') {
+        return require.resolve(`${cwd}/${path}`);
+    }
+    return require.resolve(`babel-${type}-${path}`);
+}
+
+
 if (options.babel) {
     const babelRegister = require('babel-register');
     const config = {
-        presets: [
-            require.resolve('babel-preset-latest'),
-        ],
-        plugins: [],
+        presets: options.babelPresets.map(item =>
+            resolveBabelExt('preset', item)),
+        plugins: options.babelPlugins.map(item =>
+            resolveBabelExt('plugin', item)),
     };
     if (options.coverage) {
         config.plugins.push([
@@ -36,13 +73,14 @@ if (options.babel) {
                 exclude: [
                     `**/${options.testDirectory}/**/*`,
                     '**/node_modules/**/*'
-                ],
+                ].concat(options.coverageExclude),
             },
         ]);
     }
     for (let i = 0; i < plugins.length; i++) {
         plugins[i].initBabel(config);
     }
+    config.plugins.push(require.resolve('./babel-hoist'));
     babelRegister(config);
 }
 
@@ -51,7 +89,11 @@ registerRequireHook();
 
 
 for (let i = 0; i < plugins.length; i++) {
-    plugins[i].init(options);
+    plugins[i].init({
+        options,
+        getCurrentTest,
+        events
+    });
 }
 
 
@@ -60,7 +102,8 @@ runAllTest().then(() => {
         console.log('Done');
         watch();
     } else if (options.coverage) {
-        coverage(() => console.log('Done'));
+        coverage();
+        console.log('Done');
     } else {
         console.log('Done');
     }