+++ src/require-hook.js
@@ -3,9 +3,8 @@ import Module from 'module';
 import { mockObject } from './automock';
 import { setTestFile, addDependency, getMock } from './mokamok';
 import { options } from './config';
-import { cleanUp as cleanUpAutomock, reset as resetAutomock } from './automock';
+import { cleanUp as cleanUpAutomock } from './automock';
 import { cleanUp as cleanUpMokamok } from './mokamok';
-import { cleanUp as cleanUpRequireHook } from './require-hook';
 
 
 const rootPath = path.resolve('.');
@@ -16,8 +15,20 @@ let _load = null;
 let cache = {};
 let deps = {};
 
+const mockedModules = [];
+
+const ignore = {
+    '.css': true,
+    '.scss': true,
+    '.sass': true,
+}
+
 
 function load(request, parent, isMain) {
+    if (ignore[path.extname(request)]) {
+        return {};
+    }
+
     const modulePath = Module._resolveFilename(request, parent, isMain);
 
     const isTestFile = modulePath.indexOf(testDir) !== -1;
@@ -27,6 +38,13 @@ function load(request, parent, isMain) {
         cleanUpAutomock();
         cleanUp();
         setTestFile(modulePath);
+    }
+
+    const mock = getMock(modulePath);
+
+    if (mock && mock.mock) {
+        mockedModules.push(modulePath);
+        return mock.mock;
     }
 
     let module = _load(request, parent, isMain);
@@ -39,6 +57,10 @@ function load(request, parent, isMain) {
         return cache[modulePath];
     }
 
+    if (process.env.TRACE_MOCK &&  modulePath.indexOf(process.env.TRACE_MOCK) !== -1) {
+        console.log("load:", modulePath);
+    }
+
     const isProjectFile = modulePath.indexOf(rootPath) === 0 &&
         modulePath.indexOf('/node_modules/') === -1 && !isTestFile;
 
@@ -48,21 +70,22 @@ function load(request, parent, isMain) {
         dep.push(modulePath);
     }
 
-    const mock = getMock(modulePath);
-
     if (mock === false) {
+        if (process.env.TRACE_MOCK &&  modulePath.indexOf(process.env.TRACE_MOCK) !== -1) {
+            console.log("unmock:", modulePath);
+        }
+        mockedModules.push(modulePath);
         return module;
     }
 
     const autoMock = options.automock && isProjectFile;
 
     if (mock || autoMock) {
-        if (mock  && mock.mock) {
-            module = mock.mock;
-        } else {
-            module = mockObject(module);
+        if (process.env.TRACE_MOCK &&  modulePath.indexOf(process.env.TRACE_MOCK) !== -1) {
+            console.log("mock:", modulePath);
         }
-        cache[modulePath] = module;
+        mockedModules.push(modulePath);
+        module = cache[modulePath] = mockObject(module);
     }
 
     return module;
@@ -77,13 +100,20 @@ export function registerRequireHook() {
 }
 
 
-export function uncache(name) {
+export function uncache(name, done = {}) {
+    if (process.env.TRACE_MOCK &&  name.indexOf(process.env.TRACE_MOCK) !== -1) {
+        console.log("uncache:", name);
+    }
+    done[name] = true;
     delete cache[name];
     delete Module._cache[name];
     const dep = deps[name];
     if (dep) {
         for (let i = 0; i < dep.length; i++) {
-            uncache(dep[i]);
+            const dname = dep[i];
+            if (!done[dname]) {
+                uncache(dname, done);
+            }
         }
     }
     delete deps[name];
@@ -91,6 +121,6 @@ export function uncache(name) {
 
 
 export function cleanUp() {
-    cache = {};
-    deps = {};
+    mockedModules.forEach(uncache);
+    mockedModules.length = 0;
 }